creating a separated file to deal with concept loading
