+++ src/pages/patient/onboarding/UserInfoForm2.tsx
@@ -2,8 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Button } from '@/components/forms/button';
 import { TextField } from '@/components/forms/text_input';
 import { SelectField } from '@/components/forms/select_input';
-// import type { LocationRetrieve } from '@/api/models/LocationRetrieve';
-import { ConceptService } from '@/api/services/ConceptService';
+import {useLocationConcepts} from '@/utils/conceptLoader';
 
 // Define form data type that extends Location with form-specific fields
 export interface AddressFormData extends Partial<Location> {
@@ -29,6 +28,8 @@ interface FormErrors {
 }
 
 export function UserInfoForm2({onSubmit}: {onSubmit: (data: AddressFormData) => void }): JSX.Element {
+  const {stateOptions,  isLoading, error: conceptError} = useLocationConcepts();
+  
   const [formData, setFormData] = useState<AddressFormData>({
     zip: '',
     state: '',
@@ -113,51 +114,6 @@ export function UserInfoForm2({onSubmit}: {onSubmit: (data: AddressFormData) => 
     value: string | number;
     label: string;
   }
-  
-  const [stateOptions, setStateOptions] = useState<SelectOption[]>([]);
-  const [isLoadingConcepts, setIsLoadingConcepts] = useState(true);
-  const [conceptError, setConceptError] = useState<string | null>(null);
-
-  useEffect(() => {
-      const fetchConcepts = async () => {
-        setIsLoadingConcepts(true);
-        setConceptError(null);
-        
-        try {
-            // Brazilian states
-            const concepts = await ConceptService.apiConceptList('Brazil States', 'pt');
-            
-            // Filter concepts by their class
-            const stateConcepts = concepts.filter(concept => 
-              concept.concept_class === 'Brazil States'
-            );
-            
-            const stateOptions = stateConcepts
-                .filter(concept => concept.concept_name != null)
-                .map(concept => ({
-                    value: concept.concept_id,
-                    label: concept.translated_name as string 
-                }));
-          setStateOptions(stateOptions);
-          // Log the fetched data for debugging
-         console.log('Fetched state concepts:', stateConcepts);
-        }
-        catch (error) {
-          console.error('Error fetching concepts:', error);
-          setConceptError('Erro ao carregar opções do servidor. Tente novamente mais tarde.');
-          // Fallback to hardcoded options if fetch fails
-          setStateOptions([
-            { value: "AC", label: "AC" },
-            { value: "SP", label: "SP" }
-          ]);
-        }
-        finally {
-          setIsLoadingConcepts(false);
-        }
-      };
-          
-    fetchConcepts();
-  }, []);
 
   return (
     <form onSubmit={handleSubmit} className="flex flex-col gap-4" noValidate>
